requirements.txt
azure-functions==1.18.0
azure-functions-openapi==0.5.0
azure-data-tables==12.7.0
azure-core==1.30.2
PyJWT[crypto]==2.9.0
requests==2.32.3
pytest==8.3.3
debugpy==1.8.17
